+++ src/Symfony/Component/PropertyAccess/PropertyPathInterface.php
@@ -16,6 +16,8 @@ namespace Symfony\Component\PropertyAccess;
  *
  * @author Bernhard Schussek <bschussek@gmail.com>
  *
+ * @method bool isNullSafe(int $index) Returns whether the element at the given index is null safe. Not implementing it is deprecated since Symfony 6.2
+ *
  * @extends \Traversable<int, string>
  */
 interface PropertyPathInterface extends \Traversable